Intorduce badgeeditable
Prior to this change, I wanted a way to enter and display content in
badges in the same form control.

This change introduces a package that exports the BadgeEditable
constructor and unobtrusively initialises any element with the
badgeeditable attribute.  The BadgeEditable constructor is intended to
be applied to an element with the Bootstrap form-control CSS class, but
the constructor will add the contenteditable attribute.  Any text
entered into the control will be wrapped with spans with the badge CSS
class.  Separate badges can be entered with a comma.  A demo page and
styles are provided.

There are bugs, some known and more to be discovered.

Signed-off-by: Corin Lawson <corin@responsight.com>