Chrome compatibility
Prior to this change, Chrome (and Chromium based) browsers appear to
handle Selection#collapse differently to Firefox.  The cursor appears to
snap back immediately after the first non-whitespace character
preceeding the specified offset within a certain distance.  I observed
that when the selection is set to the first position in an empty span
element the cursor would appear and insert text to the end of the
previous span instead of the span specified to the collapse method.

This change works around the issue by inserting a superflous span.  The
issue is detected by temporarily adding a div to the body and testing
the nodes in the selection after performing the problematic collapse
operation.