Use modern naming conventions
Prior to this change, ES6 module was being written without marking it as
such in the file name or package.json.

This change renames index.js to the mjs file extension and adds the
module key to package.json; the main key now points to the js file built
by rollup.