Drop non-core dependencies
Prior to this change, some dependencies were not used others were no
essential to the library operation.

This change removes bootstrap and sources it from the CDN in the demo
html file.  The dependency on onload-function-stack is dropped because
it doesn't spark joy.  The babel dependencies weren't being used.
Finally the compilation of the css file has been moved out side of
rollup.