Introduce customisable parsing and labelling
Prior to this change, input was split on commas only and resulting
badges were assigned the badge-primary class name.

This change allows the user to specify a custom parser for spliting up
the input into badges and a custom label for the class name; both
default to the previous behaviour.

The parser passed to the constructor should have a `parse` method that
accepts the text of the badge currently being entered and returns an
array of values of any type.  The array may contain undefined values,
furthermore if the array ends with an undefined value then this
signifies that the entered badge is complete and the control begins a
new badge.  Undefined values other than at then end of the array are
ingored.  More than one non-undefined values may be returned and the
current badge will be split into two.  More than two non-undefined
values are not currently supported.