Remove SNAP feature detection
Prior to this change, a cross-browser compatibility issue was addressed
by detecting a certain feature with the SNAP constant.  It turns out
that the issue was actually caused by a missing BR element in the
creation/manipulation of some badges.  Adding the BR element if it is
missing means we don't need the SNAP workaround.

Everytime a badge is enabled, this change ensures that it always has an
empty badge to the right and left and also ensures that it contains the
sentinel BR element.  This allows us to remove the SNAP feature
detection and means that splitting a badge that is already enabled works
as expected.