Pass the offset through to collapse
Prior to this change, requesting collapse during badge activation always
collapsed to the start.  It is expected that we will need to specify the
offset in future.  Also, it was found that the cursor would be placed at
the end of the input when the control first gains focus instead of the
badge/position that was clicked on.

This change passes the collapse offset through activateBadge's arguments
and introduces a small delay when activating the badge that receives the
initial focus.