Introduce change callback
Prior to this change, the component didn't communicate changes to the
user of BadgeEditable.

This change introduces value property to the BadgeEditable instance and
a onChange callback to the constructor's options for communicating any
changes to the current value of the input.  The callback mechanism is
rudimentry and likely to change.