Refactor structure and names of instance members
Prior to this change, the textContent of the badge nodes was added to
the value produced by the parser.  We have no control or responsibility
over this value and it seems to be in poor taste to be adding/changing
it.

This change stores the textContent beside the value from the parse in a
new structure that is stored in the badgeMap.  Also some names have
changed to be more consistent with their conceptual value.