Replace badge text with sanitised text
Prior to this change, entering a character that resulted in more than
one values from the parser would result in that character being omitted
from the badge text.  Sometimes this character is not a comma!

This change replaces the badge's textContent with the text property of
the value produced by the parser, otherwise the textContent is
constructed from the original input text and location offsets provided
by the parser.  If the parser doesn't provide either text or location
information then we fallback to the previous behaviour.  Careful
attention has been given to keeping the selection position the same
after the new badges have been added, however, if the text of the output
of the parser is different to its input (e.g. the parser collapses
whitespace) then results may vary.  To this end, the default parser
provides both the text and location information.

Also, we ensure that valid badges do not have the badge-empty class and
we now use the text node of the badge when collapsing the selection.

Signed-off-by: Corin Lawson <corin@responsight.com>