Provide custom sentinal hook
Prior to this change, the sentinal of each badge was alway a br element.

This change allows the parser to provide a custom sentinal per badge by
either a function or a DOM node in the sentinal property of the parser
output.  Empty badges will still be created with a br element but it
will be replaced once the badge becomes valid.

Signed-off-by: Corin Lawson <corin@responsight.com>