Ensure badgeMap keys are consistent type
Prior to this change, a Number or String was being used as the key into
badgeMap depending on where the key came from (i.e. from
badgeKeySequence or an element's dataset).  Maps work best with
consistently typed keys.

This change converts keys to Number when extracting from an element's
dataset and converts keys to a String when storing to a dataset.